test/connectivity: add individual options for netperf test types
This commit adds options to configure with netperf tests are
executed during a performance test run, allowing for the
suite of executed tests to be customized depending on the use
case. Prior, a default suite of TCP and UDP stream/RR tests were
executed, with CRR being able to be specified through a separate
option. This is important because it lets users specifically craft
the netperf tests they need to execute, without having to wait for
other tests they don't need to finish up before getting results.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>